correction via, Bob
"One minor correction I would ask that you make: I am not the inventor of the MobRetreat workshop format. At best I can identify with being a co-inventor. What was created was from a collaboration. The primary creator who deserves credit for coming up with the idea in the first place is
@dimimuc"